requirements.txt
streamlit
PyPDF2
PyMuPDF
pillow
